elevate the recentSessions state to app level, similar to investList state.
